Initial Upload
Uploaded the current version of the implementation. Currently includes only the architecture of the VGG16 classifier and functions to facilitate training. Dataset can be downloaded at the link in the ReadMe. The network is too large for my personal computer so I will attempt running on Virginia Tech's ARC clusters. If training from scratch is not an option, transfer learning will be explored.